+++ NEWS.md
@@ -1,3 +1,13 @@
+# spant 1.1.0
+* Improved ppm labels for ABfit plot results.
+* Bug fix for plot_slice_fit when using fits from masked data.
+* Updated unit tests.
+* Added vignette on manually adjusting ABfit baseline smoothness.
+* Updated the introduction vignette.
+* Bug fixes for image and stackplot functions for masked MRS data.
+* Bug fix for using RATS with masked MRS data.
+* Changes to prepare for for R 4.0.0.
+
 # spant 1.0.0
 * ABfit analysis method has been added, and is now default for the mrs_fit
 function.